backend/pyproject.toml
[project]
name = "spraply-fastapi"
version = "0.1.0"
description = "FastAPI implementation of spraply backend API"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "alembic>=1.16.4",
    "psycopg[binary]>=3.2.9",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
